fix!: Remove label:LANG_CODE style properties
BREAKING CHANGE: The schema has always defined the label, name and
placeholder properties as also supporting a prefix of `:LANG_CODE` where
LANG_CODE is the two-letter ISO language code.
This has never been used in any of the Mapeo apps, and the current
approach to translation is to load translations at runtime and keep
them separate to the preset definitions themselves.
This is a breaking change because it removes properties from the schema,
but for Mapeo Mobile and Desktop this should not break anything because
these properties were not used anyway.