feat: replaced protobuf-es with ts-proto
`protobuf-es` generated a class where each field that is not optional has
a default value (i.e. string -> '') which can't be changes. This doesn't
seem like a useful way of preventing errors since an empty object is a
valid initializer.
`ts-proto`  generates typescript types and actually enforce type checking
and its easy to integrate with `buf`
Additionaly ts-proto provides decode and encode functions so maybe
wrapping in a class is not necessary anymore?