feat: use json schemas directly
I've added json-schema-to-typescript to turn all the schemas to .d.ts
files. This replaces typebox use since it seemed easier.
Additionally this meant setting the ts-proto opt to turn snake_case
fields of protobuf schema to camelCase to false since fields in json
schema are snake_case.
Jsonschema treats dates as strings with a special format, so every proto
field that was a Timestamp now is just a string (for now)