feat: add test
Additionally I've modified the tsconfig since we were emitting only
declaration files and not transpiling to js. This disallows importing
protobuf schemas from js (after turned to .ts using ts-proto).
I set allowJs to false so that we don't transpile `src/index.js` but we
generate a declaration file for it