chore: added ts-index npm script
To use protobufs types generated from ts-proto we need to transpile ts
to js, not only emit declaration files.
To generate documentation for the index.js, we need to generate
a declaration file for index.js.
This means that `allowJs:true`, and `emitDeclarationOnly:false`. The
problem with that is that transpiling a js file would overrite it and
tsc complains about that.
So, I created a separate script to only generate a .d.ts from index.js.
Not the best solution, but works for now