feat: handle type casing with schemaPrefix
Casing on a doc type shouldn't be handled specifically by the code. Nevertheless, some schemas enforce a type to a specific value (type must be i.e. 'observation', which is different from 'Observation'). This means that type "observation"  and "Observation" should have different dataTypeIds. That way we know what "type" to fill-in on returning the decoded object to be validated...
If we want the same `dataTypeId` for 'observation' and 'Observation'
(since we know we're actually talking about roughly the same type of
'document'), we could format the schemasPrefix object differently like:
Observation: `{dataTypeId: '924892', versions: {'observation': 4, 'Observation': 5}}`
or something less worse...