feat: export fns for versionId, seq->index... (#100)
coreId->coreKey

Exports functions for parsing a string `versionId` to get the core key
and index, and for generating a string versionId. Also renames the
parameters: The naming convention of `seq` goes back many years to when
we used Hyperlog (precursor to Hypercore). `index` is the naming
currently used in Hypercore docs and I think it makes the most sense -
it refers to the index of the block within the hypercore. We also have a
convention in Mapeo Core of using `key` to refer to a Buffer, and `id`
to refer to a string. `docId` is the exception when named on the
protobuf, because it doesn't make sense to rename it when decoding, and
because `docId` as a string is never exposed outside this library.