Make field.options optional
At some point we will make the JSON schema for field a discriminate
union based on the `type` field, and `options` will only be valid for
the select one / select many types. It doesn't make sense to turn this
into and empty array when it is undefined when we do the encode-decode
round-trip. This does mean that setting it as an empty array will return
`undefined` when decoding, but this should be an array of length at
least one when defined anyway - a future schema update should enforce
this.