fix: don't set default values for optional fields (#117)
* feat: add tests for valid docs and badDocName

* feat: added test for deep comparison of fields of doc with decodedDoc

* feat: fix tests for nested objects (one-level) and floats

when decoding a protobuf with a float, the number of decimals is not
predictable nor it will necessary match the number of decimals on the
field we encoded previously...

* chore: refactor ifelse to switch

* feat: fix bug in `encode-convertions/convertProject`, refactor tests

* feat: observation.proto/{lat,lon}: use double instead of float

* add fields with default value to `field` to pass deepEqual test

* feat: add tests for optional values, proto and schema changes

* feat: add configStore tests with additional fields

* feat: fix test after master merge

* feat: add tests for extra non required values, and cleanup

* feat: fix buf in encoding field tags

* chore: format with prettier

* remove unused import

* Make badDocs test more DRY

* remove unused file

* move fixtures into separate files

* fix typos and make goodDocs correct types

* fix typo acurracy -> accuracy

* remove extra fields fixtures - not necessary

* remove doesNotThrow check

- test will fail anyway if it throws, and that's not what we're testing
 here

* only define fields that are expected to change

Re-writing the whole doc as `expected` is subject to typos, and it's
not clear what we are actually testing. Switch to just defining fields
expected to change in `expected` docs

* let -> const

* fix: don't set default values for optional fields

Fixes #116

* Make field.options optional

At some point we will make the JSON schema for field a discriminate
union based on the `type` field, and `options` will only be valid for
the select one / select many types. It doesn't make sense to turn this
into and empty array when it is undefined when we do the encode-decode
round-trip. This does mean that setting it as an empty array will return
`undefined` when decoding, but this should be an array of length at
least one when defined anyway - a future schema update should enforce
this.

* trigger CI

---------

Co-authored-by: Tomás Ciccola <tciccola@digital-democracy.com>