chore: remove `ts-ignore` and `ts-expect-error`
This change should have no user impact.

There's almost always something better than `// @ts-ignore` and
`// @ts-expect-error`. They can be problematic because they're very
broad, and typically broader than we want. For details, see
["At least 7 reasons to avoid @ts-expect-error and @ts-ignore"][0].

This removes all of them in favor of more targeted solutions. (Or, in
one case, just removes it and everything was fine!)

I think this is a useful change on its own, but it should make [an
upcoming change][1] easier too.

[0]: https://shane-o.dev/articles/any-or-expect
[1]: https://github.com/digidem/mapeo-schema/pull/243