fix!: property decode new `oneof` cases
We have proto definitions that use `oneof`, like this:

```protobuf
message Foo {
  oneof something {
    string bar = 1;
    string baz = 2;
  }
}
```

We also have decoding that does something like this:

```javascript
switch (foo.something?.$case) {
  case 'bar':
    // ...do something...
  case 'baz':
    // ...do something...
  default:
    throw new Error('Unexpected case')
}
```

This will break if we add another case, like this:

```diff
 message Foo {
   oneof something {
     string bar = 1;
     string baz = 2;
+    string qux = 3;
   }
 }
```

This fixes those cases in the codebase, treating an unknown `oneof` as
skipped. We already did this in a few places.