feat: decode with coreDiscoveryId or coreDiscoveryKey
Allow decode to work when a coreDiscoveryId is
passed, instead of coreDiscoveryKey to match an
upstream change in the multi-core-indexer to use
coreDiscoveryId rather than coreDiscoveryKey. This
reduces a round-trip conversion from buffer to
string, and will also facilitate sending entries
between threads when we run indexing in a separate
worker thread.